Add submodule with customized geth source
Inline geth nix derivation in repo and use submodule for code

We can't simply override the `src` attribute of the go-ethereum
derivation because this doesn't work with vendored go dependencies.
For more info: https://github.com/NixOS/nixpkgs/issues/86349 .

Here we copy the nix derivation from nixpkgs to ./nix/go-ethereum folder
of this repo. Nix will print a mismatch to the console if the vendored
geth dependencies change. In that case the new vendorSha256 in the
inlined geth nix derivation in ./nix/go-ethereum/ needs to be updated.

This is a stop-gap solution so we can all work with the same version
of geth until we find a better way to enable that.