Hardhat project at top level, merge package.json
I think most code in this repo will be in some way dependent on the
smart contract so I think it makes sense to move the hardhat project to
the top level directory for this initial phase. It also has the
advantage of not having to write documentation for the setup of multiple
subprojects.