Several improvements:
1) Incorrect number of inputs/outputs in "Dummy Verifier"

PROBLEM: the current implementation of the "Dummy Verifier" assumes there are only one input and output for each
AAP transaction.

SOLUTION: Set N_INPUTS=4 and N_OUTPUTS=5  based on
existing benchmarks: https://www.notion.so/translucence/Benchmark-Transfer-Note-Generation-Verification-4e3878797c8643069d1f36d0bec56e7c

2) PROBLEM: The AAP input used for tests contains a list of equal values.
It is not clear if this can affect the gas cost for Arbitrum due
to compression.

SOLUTION: Make the AAP input for tests and benchmarks look random.

3) FIX: wrong computation of gas

PROBLEM: for the benchmarks and the tests we compute the gas spent for the whole block, instead of only the transaction.