Add contract block height, check note expiry
Due to way we handle the publishing of nullifiers this introduces a
requirement that the expiry is checked *before* trying to publish the
nullifiers.

If the nullifiers are published the transaction must be included in the
block *or* the entire Ethereum transaction must be reverted.

We may want to rework this part in the future.

Close #146
Close #91