Integration of wallet with EQS (#621)
* Use real EQS event stream for CapeWalletBackend

* Use real EQS for synchronous queries, removing mock_eqs from CapeBackend

* Use real backend in CLI

* Don't let ethers filter refetch previously processed events

* Fix slow tests build

* Fix default URLs (no https)

* Integrate real backend with wallet server

* ETH polling fixes

* Change (last_block, last_index) to (next_block, next_index) so
  (0, 0) makes sense as the initial state
* Fix the check for duplicate events (previously it would sometimes
  discard events that were actually new)
* Persist pending_commit. This allows us to always use the index
  information in the persistent state and remove the duplicate
  indices from the EthPolling state.
* Use the CAPE block ID for transaction maps, not the ETH block ID.

* Improved model of event indexing for EQS

* Fix web server tests

I moved the web server related files out of the lib target and into
the web_server bin target, so that #[cfg(test)] can be used to select
between the mock backend and the real backend. This makes it harder to
accidentally select the mock backend then using feature flags. I had
dreams of making the whole web server generic on the backend type, but
alas, this triggers an ICE.

Also moved the web server bin from wallet/src/bin to wallet/src/web_server,
since module files cannot be placed in bin/ or accessed from a binary if
they are placed in src/.

* Implement review feedback

* Build events before locking state in eth_polling
* Make wallet polling frequency a parameter
* Make address book URL used by wallet a parameter

Co-authored-by: Nathan F Yospe <nyospe@gmail.com>