Asynchronous faucet. (#1157)
* Asynchronous faucet.

Instead of processing requests directly in `tide` request handlers,
the request handler simply adds the request to a queue, and the
requests are dequeued and processed by a fixed number of worker
threads.

This solves two problems:
* Requests taking a long time to process do not time out the HTTP
  request
* We limit the CPU usage by limiting the number of worker threads,
  so the server does not get overloaded.

Closes #1151

* Make the faucet's queue persistent, so we don't lose requests if the faucet crashes

* Test faucet restartability

* Persist number of grants given to each key.

This fixes a race condition where we could be shut down in the middle
of transferring several grants to a key. After restarting, previously,
we would start granting to this key from scratch, potentially granting
more than we intended and even, if we get shut down frequently, not
making progress at all.

Note, there is still a small race condition: if we get shut down just
after building a transfer but before incrementing the count of transfers
to that key (a very small window of time) we may repeat the grant the
next time we are restarted. This is a minor problem, as it means at
worst tranfserring an extra grant per restart. To fix this, we need
to synchronize between the faucet's storage and the wallet library
storage, which I think requires some kind of composabilities between
AtomicStores.