Arbitrum tests (#1230)
* Run all tests against arbitrum
* Higher gas limit override for arbitrum
* Remove hardcoded provider URL

* Ensure non-deployer is not deployer

If `CAPE_WEB3_PROVIDER_URL` is `get_funded_client()` returns the same
client each time its called.

Use a `call` instead of a `send` to avoid requiring a funded client.

This still ensures that the call fails for the right reason (the
function was called from the wrong account) because we match against the
revert reason.

* Enable running wallet tests on arbitrum
* Update docs for arbitrum testing
* relayer: add extra gas to transaction

We previously set the limit manually but this is tricky because gas cost
is different on arbitrum. By adding the 10M to the estimate instead it
should work for both L1's and arbitrum.

Remove the test that tests if the extra gas has effect. The test would
be a bit involved to rewrite right now because we can't set a too low
gas limit directly anymore.

* Add 1M extra gas for tx that sometimes runs out